Initial rewrite of the stack
We have decided to swap out Django for FastAPI as our base.
This does give us more work, as a lot of things were handled by Django, which we have to now do our selves.
We do however feel that this change is for the better as we were fighting against Django in various places.

We also started using `poetry`for dependency management in Python

For the frontend we started rebuilding the entire UI using Nuxt 3 and Nuxt UI 2